Move rust/python logic to python crate
