+++ erdos/src/lib.rs
@@ -115,7 +115,6 @@
 
 // Required for specialization.
 #![allow(incomplete_features)]
-#![feature(get_mut_unchecked)]
 #![feature(specialization)]
 #![feature(box_into_pin)]
 
@@ -134,8 +133,6 @@ use uuid;
 
 // Private submodules
 mod configuration;
-#[cfg(feature = "python")]
-mod python;
 
 // Public submodules
 #[doc(hidden)]
