+++ python/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2018"
 [dependencies]
 erdos = { path = "../erdos" }
 pyo3 = { version = "0.15.1", features = ["extension-module"] }
+tracing = "0.1.29"
 
 [lib]
 name = "erdos"
