+++ python/src/py_operators/py_two_in_one_out.rs
@@ -1,17 +1,16 @@
 use std::{mem, sync::Arc};
 
-use crate::{
-    dataflow::{
-        context::OneInOneOutContext,
-        operator::{OneInOneOut, OperatorConfig},
-        stream::WriteStreamT,
-        Message, ReadStream, WriteStream,
-    },
-    python::{PyReadStream, PyTimestamp, PyWriteStream},
+use erdos::dataflow::{
+    context::TwoInOneOutContext,
+    operator::{OperatorConfig, TwoInOneOut},
+    stream::WriteStreamT,
+    Message, ReadStream, WriteStream,
 };
 use pyo3::{prelude::*, types::*};
 
-pub(crate) struct PyOneInOneOut {
+use crate::{PyReadStream, PyTimestamp, PyWriteStream};
+
+pub(crate) struct PyTwoInOneOut {
     py_operator_config: Arc<PyObject>,
     py_operator: Arc<PyObject>,
     // The py_write_stream is set to Option, since the constructor does not receive a WriteStream,
@@ -19,7 +18,7 @@ pub(crate) struct PyOneInOneOut {
     py_write_stream: Option<Arc<PyObject>>,
 }
 
-impl PyOneInOneOut {
+impl PyTwoInOneOut {
     pub(crate) fn new(
         py_operator_type: Arc<PyObject>,
         py_operator_args: Arc<PyObject>,
@@ -46,10 +45,11 @@ impl PyOneInOneOut {
     }
 }
 
-impl OneInOneOut<(), Vec<u8>, Vec<u8>> for PyOneInOneOut {
+impl TwoInOneOut<(), Vec<u8>, Vec<u8>, Vec<u8>> for PyTwoInOneOut {
     fn run(
         &mut self,
-        read_stream: &mut ReadStream<Vec<u8>>,
+        left_read_stream: &mut ReadStream<Vec<u8>>,
+        right_read_stream: &mut ReadStream<Vec<u8>>,
         write_stream: &mut WriteStream<Vec<u8>>,
     ) {
         // Note on the unsafe execution: To conform to the Operator API that passes a mutable
@@ -61,15 +61,22 @@ impl OneInOneOut<(), Vec<u8>, Vec<u8>> for PyOneInOneOut {
         // since pyo3 does not support annotation of pyclasses with lifetime parameters.
         unsafe {
             // Create another object that points to the same original memory location of the
-            // ReadStream.
-            let read_stream_ptr: *mut ReadStream<Vec<u8>> = read_stream;
-            let read_stream: ReadStream<Vec<u8>> = read_stream_ptr.read();
+            // first ReadStream, and create the Python version of the ReadStream.
+            let left_read_stream_ptr: *mut ReadStream<Vec<u8>> = left_read_stream;
+            let left_read_stream: ReadStream<Vec<u8>> = left_read_stream_ptr.read();
+            let left_read_stream_id = left_read_stream.id();
+            let left_read_stream_name = String::from(left_read_stream.name().clone());
+            let left_read_stream_arc = Arc::new(left_read_stream);
+            let py_left_read_stream = PyReadStream::from(&left_read_stream_arc);
 
-            // Create the Python version of the ReadStream.
-            let read_stream_id = read_stream.id();
-            let read_stream_name = String::from(read_stream.name().clone());
-            let read_stream_arc = Arc::new(read_stream);
-            let py_read_stream = PyReadStream::from(&read_stream_arc);
+            // Create another object that points to the same original memory location of the
+            // second ReadStream, and create the Python version of the ReadStream.
+            let right_read_stream_ptr: *mut ReadStream<Vec<u8>> = right_read_stream;
+            let right_read_stream: ReadStream<Vec<u8>> = right_read_stream_ptr.read();
+            let right_read_stream_id = right_read_stream.id();
+            let right_read_stream_name = String::from(right_read_stream.name().clone());
+            let right_read_stream_arc = Arc::new(right_read_stream);
+            let py_right_read_stream = PyReadStream::from(&right_read_stream_arc);
 
             // Create the Python version of the WriteStream.
             let write_stream_clone = write_stream.clone();
@@ -78,19 +85,43 @@ impl OneInOneOut<(), Vec<u8>, Vec<u8>> for PyOneInOneOut {
             let py_write_stream = PyWriteStream::from(write_stream_clone);
 
             // Create the locals to run the constructor for the ReadStream and WriteStream.
-            let (py_read_stream_obj, py_write_stream_obj) =
-                Python::with_gil(|py| -> (PyObject, PyObject) {
+            let (py_left_read_stream_obj, py_right_read_stream_obj, py_write_stream_obj) =
+                Python::with_gil(|py| -> (PyObject, PyObject, PyObject) {
                     let locals = PyDict::new(py);
                     locals
-                        .set_item("py_read_stream", &Py::new(py, py_read_stream).unwrap())
+                        .set_item(
+                            "py_left_read_stream",
+                            &Py::new(py, py_left_read_stream).unwrap(),
+                        )
                         .err()
                         .map(|e| e.print(py));
                     locals
-                        .set_item("read_stream_id", format!("{}", read_stream_id))
+                        .set_item("left_read_stream_id", format!("{}", left_read_stream_id))
                         .err()
                         .map(|e| e.print(py));
                     locals
-                        .set_item("read_stream_name", format!("{}", read_stream_name))
+                        .set_item(
+                            "left_read_stream_name",
+                            format!("{}", left_read_stream_name),
+                        )
+                        .err()
+                        .map(|e| e.print(py));
+                    locals
+                        .set_item(
+                            "py_right_read_stream",
+                            &Py::new(py, py_right_read_stream).unwrap(),
+                        )
+                        .err()
+                        .map(|e| e.print(py));
+                    locals
+                        .set_item("right_read_stream_id", format!("{}", right_read_stream_id))
+                        .err()
+                        .map(|e| e.print(py));
+                    locals
+                        .set_item(
+                            "right_read_stream_name",
+                            format!("{}", right_read_stream_name),
+                        )
                         .err()
                         .map(|e| e.print(py));
                     locals
@@ -109,8 +140,11 @@ impl OneInOneOut<(), Vec<u8>, Vec<u8>> for PyOneInOneOut {
                         r#"
 import uuid, erdos
 
-# Create the ReadStream.
-read_stream = erdos.ReadStream(_py_read_stream=py_read_stream)
+# Create the first ReadStream.
+left_read_stream = erdos.ReadStream(_py_read_stream=py_left_read_stream)
+
+# Create the first ReadStream.
+right_read_stream = erdos.ReadStream(_py_read_stream=py_right_read_stream)
 
 # Create the WriteStream.
 write_stream = erdos.WriteStream(_py_write_stream=py_write_stream)
@@ -123,8 +157,12 @@ write_stream = erdos.WriteStream(_py_write_stream=py_write_stream)
                     }
 
                     // Retrieve the constructed stream.
-                    let py_read_stream_obj = py
-                        .eval("read_stream", None, Some(&locals))
+                    let py_left_read_stream_obj = py
+                        .eval("left_read_stream", None, Some(&locals))
+                        .unwrap()
+                        .to_object(py);
+                    let py_right_read_stream_obj = py
+                        .eval("right_read_stream", None, Some(&locals))
                         .unwrap()
                         .to_object(py);
                     let py_write_stream_obj = py
@@ -132,7 +170,11 @@ write_stream = erdos.WriteStream(_py_write_stream=py_write_stream)
                         .unwrap()
                         .to_object(py);
 
-                    (py_read_stream_obj, py_write_stream_obj)
+                    (
+                        py_left_read_stream_obj,
+                        py_right_read_stream_obj,
+                        py_write_stream_obj,
+                    )
                 });
 
             // Save the constructed WriteStream, and invoke the `run` method.
@@ -143,7 +185,11 @@ write_stream = erdos.WriteStream(_py_write_stream=py_write_stream)
                 if let Err(e) = self.py_operator.call_method1(
                     py,
                     "run",
-                    (py_read_stream_obj, py_write_stream_arc_clone.clone_ref(py)),
+                    (
+                        py_left_read_stream_obj,
+                        py_right_read_stream_obj,
+                        py_write_stream_arc_clone.clone_ref(py),
+                    ),
                 ) {
                     e.print(py);
                 }
@@ -152,7 +198,8 @@ write_stream = erdos.WriteStream(_py_write_stream=py_write_stream)
             // NOTE: We must forget the Arc<ReadStream> instance since Rust calls the Drop trait
             // at the end of this block, which leads to a dropped ReadStream for the executor, thus
             // preventing the execution of the message and watermark callbacks.
-            mem::forget(read_stream_arc);
+            mem::forget(left_read_stream_arc);
+            mem::forget(right_read_stream_arc);
         }
     }
 
@@ -164,13 +211,14 @@ write_stream = erdos.WriteStream(_py_write_stream=py_write_stream)
         });
     }
 
-    fn on_data(&mut self, ctx: &mut OneInOneOutContext<(), Vec<u8>>, data: &Vec<u8>) {
+    fn on_left_data(&mut self, ctx: &mut TwoInOneOutContext<(), Vec<u8>>, data: &Vec<u8>) {
         let py_time = PyTimestamp::from(ctx.get_timestamp().clone());
         let py_operator_config = Arc::clone(&self.py_operator_config);
         let py_write_stream = match &self.py_write_stream {
             Some(d) => Arc::clone(d),
             None => unreachable!(),
         };
+
         Python::with_gil(|py| {
             let erdos = PyModule::import(py, "erdos").unwrap();
             let pickle = PyModule::import(py, "pickle").unwrap();
@@ -178,7 +226,7 @@ write_stream = erdos.WriteStream(_py_write_stream=py_write_stream)
             let context: PyObject = erdos
                 .getattr("context")
                 .unwrap()
-                .getattr("OneInOneOutContext")
+                .getattr("TwoInOneOutContext")
                 .unwrap()
                 .call1((
                     py_time,
@@ -198,14 +246,56 @@ write_stream = erdos.WriteStream(_py_write_stream=py_write_stream)
                 .unwrap();
             if let Err(e) = self
                 .py_operator
-                .call_method1(py, "on_data", (context, py_data))
+                .call_method1(py, "on_left_data", (context, py_data))
             {
                 e.print(py);
             };
         });
     }
 
-    fn on_watermark(&mut self, ctx: &mut OneInOneOutContext<(), Vec<u8>>) {
+    fn on_right_data(&mut self, ctx: &mut TwoInOneOutContext<(), Vec<u8>>, data: &Vec<u8>) {
+        let py_time = PyTimestamp::from(ctx.get_timestamp().clone());
+        let py_operator_config = Arc::clone(&self.py_operator_config);
+        let py_write_stream = match &self.py_write_stream {
+            Some(d) => Arc::clone(d),
+            None => unreachable!(),
+        };
+
+        Python::with_gil(|py| {
+            let erdos = PyModule::import(py, "erdos").unwrap();
+            let pickle = PyModule::import(py, "pickle").unwrap();
+
+            let context: PyObject = erdos
+                .getattr("context")
+                .unwrap()
+                .getattr("TwoInOneOutContext")
+                .unwrap()
+                .call1((
+                    py_time,
+                    py_operator_config.clone_ref(py),
+                    py_write_stream.clone_ref(py),
+                ))
+                .unwrap()
+                .extract()
+                .unwrap();
+            let serialized_data = PyBytes::new(py, &data[..]);
+            let py_data: PyObject = pickle
+                .getattr("loads")
+                .unwrap()
+                .call1((serialized_data,))
+                .unwrap()
+                .extract()
+                .unwrap();
+            if let Err(e) = self
+                .py_operator
+                .call_method1(py, "on_right_data", (context, py_data))
+            {
+                e.print(py);
+            };
+        });
+    }
+
+    fn on_watermark(&mut self, ctx: &mut TwoInOneOutContext<(), Vec<u8>>) {
         let py_time = PyTimestamp::from(ctx.get_timestamp().clone());
         let py_operator_config = Arc::clone(&self.py_operator_config);
         let py_write_stream = match &self.py_write_stream {
@@ -218,7 +308,7 @@ write_stream = erdos.WriteStream(_py_write_stream=py_write_stream)
             let context: PyObject = erdos
                 .getattr("context")
                 .unwrap()
-                .getattr("OneInOneOutContext")
+                .getattr("TwoInOneOutContext")
                 .unwrap()
                 .call1((
                     py_time,
