+++ python/src/py_stream/py_loop_stream.rs
@@ -1,6 +1,5 @@
+use erdos::dataflow::{LoopStream, Stream};
 use pyo3::prelude::*;
-
-use crate::dataflow::{LoopStream, Stream};
 
 use super::PyStream;
